requirements.txt
requests
python-dotenv
pycatbox